drf/requirements.txt
algoliasearch-django>=2.0,<3.0
django
djangorestframework
pyyaml
requests
django-cors-headers 